fix: refactor and fix for comments
