+++ app/src/pages/inside/common/scenario/index.ts
