+++ app/src/pages/inside/common/scenario/scenario.tsx
